==> 1-preprocess_swt_data.qmd <==
# NOTE THAT THIS FILE IS OBSOLETE. Please use 2.1-workshop_data_visualization.R for data processing
---
title: "Preprocess SWT Raw Data 1"
format: html
editor: source
---

```{r load_libraries}

library(tidyverse)
library(readxl)
library(labelled)
```

## Load datasets
# files renamed beforehand:: 
#  "swt_2014_raw.xlsx" was "2014- Copy of De-identified v1- SWT Data and Evaluation.xlsx"  
# "swt_2015_raw.xlsx" was "2015-Copy of De-Identified v1 SWT 2015 Data - General Information Forms.xlsx"
# "swt_2016_raw.xlsx" was "2016 Copy of De-Identified v1- 2016 Data- Master List.xlsx" same for 2017 on to 2020

```{r load_datasets}
files_xlsx <- c("swt_2014_raw.xlsx", "swt_2014_raw.xlsx",
                "swt_2015_raw.xlsx", "swt_2016_raw.xlsx", "swt_2017_raw.xlsx",
                "swt_2018_raw.xlsx", "swt_2019_raw.xlsx", "swt_2020_raw.xlsx")#, "swt_2021_raw.xlsx")

sheets_names <- c("Intake Forms", "Post Workshop Surveys",
                  "Master List", "2016", "Sheet1", 
                  "2018 w Gen Info Forms", "In-Person w Gen Intake Forms", "In-person w Gen Intake Forms") # note that R ignores / in sheet names

ranges <- c("A:H", "A:D", "A:K", rep ("A:M",5))

seed_names <- c("2014_preworkshop","2014_postworkshop",
               "2015_preworkshop", "2016_preworkshop", "2017_preworkshop",
               "2018_preworkshop", "2019_preworkshop", "2020_preworkshop")

for (i in seq(length(files_xlsx))) {
  source_file<- file.path(rprojroot::find_root(rprojroot::is_rstudio_project),
                        #getwd(),
                         "Data", "raw data",
                        #"Data",
                         files_xlsx[i])
  
  assign(paste0("swt_", seed_names[i]) ,
         read_excel(source_file,
                       sheet = sheets_names[i],
                       range = cell_cols(ranges[i]),
                       col_types = "text"))
}

```

```{r check_imported_dataset}
#sanity check
glimpse(swt_2016_preworkshop)
glimpse(swt_2019_preworkshop)

```

## Modify if needed

```{r}
# reason for modification: original var name is too long. it creates an error when generating data dictionary.
swt_2016_preworkshop <- swt_2016_preworkshop %>% 
  rename(`Interested in receiving more information` = 
           `Would you be interested in receiving more information about our services?`)
swt_2017_preworkshop <- swt_2017_preworkshop %>% 
  rename(`Interested in receiving more information` = 
           `Would you be interested in receiving more information about our services?`)
swt_2018_preworkshop <- swt_2018_preworkshop %>% 
  rename(`Interested in receiving more information` = 
           `Would you be interested in receiving more information about our services?`)
  rename ("Location" = "Location if different than partner" )
swt_2019_preworkshop <- swt_2019_preworkshop %>% 
  rename(`Interested in receiving more information` = 
           `Would you be interested in receiving more information about our services?`)
  rename ("Location" = "Location if different than partner" )
swt_2020_preworkshop <- swt_2020_preworkshop %>% 
  rename(`Interested in receiving more information` = 
           `Would you be interested in receiving more information about our services?`) %>%
  rename ("Location" = "Location if different than partner" )
```

## Export

```{r export_to_csv}

files_csv <- c("swt_2014_processed_pre.csv","swt_2014_processed_post.csv",
               "swt_2015_processed_pre.csv", "swt_2016_processed_pre.csv",
               "swt_2017_processed_pre.csv", "swt_2018_processed_pre.csv",
               "swt_2019_processed_pre.csv", "swt_2020_processed_pre.csv")

swt_dfs <- c("swt_2014_preworkshop","swt_2014_postworkshop",
               "swt_2015_preworkshop", "swt_2016_preworkshop",
             "swt_2017_preworkshop", "swt_2019_preworkshop",
             "swt_2018_preworkshop", "swt_2020_preworkshop")

for (i in seq(length(swt_dfs))) {
  target_file <- file.path(rprojroot::find_root(rprojroot::is_rstudio_project),
                        #getwd(),
                         "data", files_csv[i])
  write_csv(get(swt_dfs[i]), target_file)
}

```

## Create data dictionary

```{r}
# Label tibbles

swt_2014_preworkshop_labelled <- swt_2014_preworkshop %>% 
  set_variable_labels(
    Workshop     = "Workshop type",
    Date         = "Date of the workshop",
    USERID       = "Participant's id, a participant may have multiple ids",
    `Zip code`   = "Zip code where a participant's lives",
    `Interested in receiving more information`  = "Indicator whether the participants want more information",
    `Status/FInancial Situation` = "Participant's status and/ or financial situation",
    `Age Range` = "Participant's age range",
    Comments    = "Participant's comments (if any)"
  )

swt_2014_postworkshop_labelled <- swt_2014_postworkshop %>% 
  set_variable_labels(
    `Workshop Type`    = "Workshop type",
    Date               = "Date of the workshop",
    Rating             = "Participants' rating for the workshop",
    Feedback           = "Participants' feedback"
  )


swt_2015_preworkshop_labelled <- swt_2015_preworkshop %>% 
  set_variable_labels(
    Workshop     = "Workshop type",
    Date         = "Date of the workshop",
    Location     = "Location of the workshop",
    USERID       = "Participant's id, a participant may have multiple ids",
    `Zip code`   = "Zip code where a participant's lives",
    Race         = "Participant's race",
    Gender       = "Participant's gender",
    `How did you hear about us` = "Participant’s source of information about SWT",
    `Age Range`  = "Participant's age range",
    `Interested in receiving more information` = "Indicator whether the participants want more information",
    Comments    = "Participant's comments (if any)"
  )

swt_2016_preworkshop_labelled <- swt_2016_preworkshop %>% 
  set_variable_labels(
    Workshop    = "Workshop type",
    Date        = "Date of the workshop",
    USERID      = "Participant's id, a participant may have multiple ids",
    `Zip Code:` = "Zip code where a participant's lives",
    `Race:`     = "Participant's race",
    `Gender:`   = "Participant's gender",
    `How did you hear about this workshop?` = "Participant’s source of information about the workshop",
    `Current Status (please select one):`   = "Participant's status",
    `Age Range (please select one):`        = "Participant's age range",
    `Interested in receiving more information` = "Indicator whether the participants want more information",
    `Comments / Please Help me With` = "Participant's comments or request (if any)",
    `Partner (If Applicable)`        = "Partner to run the workshop (if any)",
    Location    = "Workshop's delivery mode"
  )
```


```{r}
# generate data dictionaries

dfs_labelled <- c("swt_2014_preworkshop_labelled", "swt_2014_postworkshop_labelled",
                  "swt_2015_preworkshop_labelled", "swt_2016_preworkshop_labelled")

for (i in seq(length(dfs_labelled))) {
  seed_name <- str_sub(dfs_labelled[i], 1, -10)
  assign(paste0(seed_name, "_dictionary") ,
         get(dfs_labelled[i]) %>% generate_dictionary()
         )
}

```


```{r}
#sanity check
swt_2014_postworkshop_dictionary
```

```{r}
# export data dictionaries
dict_list <- c("swt_2014_preworkshop_dictionary",
               "swt_2014_postworkshop_dictionary",
               "swt_2015_preworkshop_dictionary",
               "swt_2016_preworkshop_dictionary")

dict_csv <- c("ddict_swt_2014_pre.csv","ddict_swt_2014_post.csv",
               "ddict_swt_2015_pre.csv", "ddict_swt_2016_pre.csv")

for (i in seq(length(dict_list))) {
  target_file <- file.path(rprojroot::find_root(rprojroot::is_rstudio_project),
                        #getwd(),
                         "data", dict_csv[i])
  write_csv(get(dict_list[i]), target_file)
}

```
